Added new camera.py module with CameraManager class with Snapshot functionalities
